Refactored to reduce repetitive code
