--- src/main/java/edu/iu/jrsalata/StatementFactoryInterface.java
@@ -1,13 +1,0 @@
-// Class: StatementFactoryInterface
-// This is an interface that will define the methods that concretions will use to create statements
-package edu.iu.jrsalata;
-
-public interface StatementFactoryInterface {
-    public Statement processStatement(String statement) throws InvalidAssemblyFileException;
-
-    public HexNum getStart();
-
-    public HexNum getLen();
-
-    public String getName();
-}
